fix: update iOS client, add visitor data to YouTube requests
Theta-Dev have fixed the issue with the iOS client returning invalid responses
Apparently the only issue with the iOS client is that we were not sending it visitor data cookies. Since the visitor data cookies are just protobuf objects with a random ID, timestamp and country code, they can be randomly generated and dont need to be fetched from YouTube.

Co-Authored-By: ThetaDev <thetadev@magenta.de>